[move]ShipTypeGroupクラスをlogbook.constantsパッケージへ移動
