+++ src/main/java/logbook/internal/gui/DeckShipPane.java
@@ -31,6 +31,7 @@ import javafx.scene.layout.VBox;
 import javafx.scene.text.TextFlow;
 import javafx.util.StringConverter;
 import logbook.bean.AppDeck.AppDeckShip;
+import logbook.constants.ShipTypeGroup;
 import logbook.bean.Ship;
 import logbook.bean.ShipCollection;
 import logbook.bean.SlotItem;
@@ -40,7 +41,6 @@ import logbook.bean.SlotitemMstCollection;
 import logbook.internal.Items;
 import logbook.internal.LoggerHolder;
 import logbook.internal.ShipFilter;
-import logbook.internal.ShipTypeGroup;
 import logbook.internal.Ships;
 import logbook.internal.ToStringConverter;
 import logbook.internal.gui.PopupSelectbox.SelectboxCell;
